requirements.txt
click==8.1.7
colorama==0.4.6
exceptiongroup==1.3.0
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
rich==14.2.0
shellingham==1.5.4
tomli==2.3.0
typer==0.12.3
typing_extensions==4.15.0
